+++ app/models/arbetsformedlingen/job_wrapper.rb
@@ -4,10 +4,10 @@ module Arbetsformedlingen
   class JobWrapper
     attr_reader :packet
 
-    def initialize(arbetsformedlingen_ad)
+    def initialize(arbetsformedlingen_ad, staffing_company: nil)
       @arbetsformedlingen_ad = arbetsformedlingen_ad
       @job = arbetsformedlingen_ad.job
-      @employer = JobEmployer.new(@job)
+      @employer = staffing_company || @job.staffing_company || @job.company
       @af_models = {}
 
       @packet = build_packet
@@ -82,9 +82,8 @@ module Arbetsformedlingen
 
     def build_company
       @af_models[:company] ||= Arbetsformedlingen::Company.new(
-        name: employer.name,
+        name: employer.display_name,
         cin: employer.formatted_cin,
-        description: employer.description,
         address: {
           country_code: employer.country_code,
           zip: employer.zip,