+++ app/models/linkedin/job_wrapper.rb
@@ -4,12 +4,14 @@ module Linkedin
   class JobWrapper
     attr_reader :job, :company
 
-    def initialize(job:)
+    def initialize(job:, staffing_company: nil)
       @job = job
-      @company = JobEmployer.new(job)
+      @company = staffing_company || job.staffing_company || job.company
     end
 
-    delegate :name, to: :company, prefix: true
+    def company_name
+      company.display_name
+    end
 
     def partner_job_id
       job.to_param