+++ app/services/find_frilans_finans_user_id_service.rb
@@ -2,7 +2,7 @@
 
 class FindFrilansFinansUserIdService
   def self.call(email:)
-    document = FrilansFinansApi::User.index(email: email)
+    document = FrilansFinansAPI::User.index(email: email)
     document.resource.id
   end
 end