+++ app/services/frilans_finans_invoice_attributes_service.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 class FrilansFinansInvoiceAttributesService
-  def self.call(user:, job:, pre_report:, express_payment:, client: FrilansFinansApi.config.client_klass.new) # rubocop:disable Metrics/LineLength
-    tax = FrilansFinansApi::Tax.index(only_standard: true, client: client).resource
+  def self.call(user:, job:, pre_report:, express_payment:, client: FrilansFinansAPI.config.client_klass.new) # rubocop:disable Metrics/LineLength
+    tax = FrilansFinansAPI::Tax.index(only_standard: true, client: client).resource
     # We need to update the users profession title to match the jobs,
     # in order to please Frilans Finans
-    ff_user = FrilansFinansApi::User.update(
+    ff_user = FrilansFinansAPI::User.update(
       id: user.frilans_finans_id!,
       attributes: { profession_title: job.category.name }
     )