+++ spec/models/job_spec.rb
@@ -105,7 +105,6 @@ RSpec.describe Job, type: :model do
     it 'contains headings for all present fields' do
       job = FactoryGirl.build(:job, applicant_description: 'Testing')
       result = job.full_standalone_description
-      expect(result).to include(I18n.t('job.description_title'))
       expect(result).to include(I18n.t('job.applicant_description_title'))
       expect(result).not_to include(I18n.t('job.tasks_description_title'))
       expect(result).not_to include(I18n.t('job.requirements_description_title'))