+++ spec/spec_helper.rb
@@ -14,13 +14,12 @@ end
 require 'webmock/rspec'
 require 'frilans_finans_api'
 require 'frilans_finans_api/test_helper'
-require 'json_api_helpers'
 require 'byebug'
 Dir['spec/spec_support/spec_helpers/**/*.rb'].each { |f| require_relative "../#{f}" }
 
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
-  config.include FrilansFinansApi::TestHelper
+  config.include FrilansFinansAPI::TestHelper
 
   config.expect_with :rspec do |expectations|
     # Better expetations output
@@ -54,12 +53,12 @@ RSpec.configure do |config|
 end
 
 # Set test defaults
-FrilansFinansApi.configure do |config|
-  config.client_klass = FrilansFinansApi::FixtureClient
+FrilansFinansAPI.configure do |config|
+  config.client_klass = FrilansFinansAPI::FixtureClient
   config.base_uri = 'https://example.com'
   config.client_id = '123456'
   config.client_secret = 'notsosecret'
-  config.event_logger = FrilansFinansApi::NilEventLogger.new
+  config.event_logger = FrilansFinansAPI::NilEventLogger.new
 end
 
 # Only allow the tests to connect to localhost and  allow codeclimate